requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow>=11.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

